glr2.cc: factor the symbol destruction
Much of the code that was added to reclaim memory in the case of
variant was adhoc: just clear the variant.  This is (i) repetitive and
tedious, and (ii) does not call the user destructor.

Instead, have yy_destroy_ invoke variant destruction, and use only
yy_destroy_ to destroy symbols.

Unfortunately currently glr2.cc makes too many copies (for instance
when shifting a token, instead of moving it, it copies it on the stack
and destroys the lookahead), so we get additional destructions, and
spurious logs.  Until we address this move/copy issue, let's adjust
the expectations.  Eventually this part should be reverted.

* data/skeletons/glr2.cc (yy_destroy_): Handle variant destruction.
Use it everywhere we destroyed variants.
* tests/actions.at, tests/calc.at, tests/glr-regression.at: Adjust.